# internal/openapi/spec.yaml
openapi: "3.0.3"
info:
  title: SWAMP API
  description: Software Assurance Marketplace — AI-powered security analysis of Git repositories.
  version: "1.0.0"
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Development server
...
paths:
...
  /api/v1/admin/backups:
    get:
      tags: [backups]
      summary: List backups
      operationId: listBackups
      security:
        - cookieAuth: []
      responses:
        "200":
          description: List of backups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Backup"
  /api/v1/admin/backups/trigger:
    post:
      tags: [backups]
      summary: Trigger a new backup
      operationId: triggerBackup
      security:
        - cookieAuth: []
      responses:
        "202":
          description: Backup started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backup"
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: swamp_session
    bearerAuth:
      type: http
      scheme: bearer
      description: API key in Authorization header
...
  schemas:
...
    Backup:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        s3_key:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        status_detail:
          type: string
        error_msg:
          type: string
        initiated_by:
          type: string
          format: uuid
        encrypted:
          type: boolean
        size_bytes:
          type: integer
          format: int64
        checksum:
          type: string
        duration_secs:
          type: integer
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time